feat: add SendString method to Context
This commit adds a `SendString` method to the `Context` struct. This method allows sending a plain text string as a response. It sets the `Content-Type` header to `text/plain; charset=utf-8` and writes the string to the response body. A corresponding unit test is also added to verify the functionality.